% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{as.data.frame.sleepsimR_allocations}
\alias{as.data.frame.sleepsimR_allocations}
\title{Convert allocations to data frame}
\usage{
\method{as.data.frame}{sleepsimR_allocations}(x, ...)
}
\arguments{
\item{x}{object of type sleepsimR_allocations}
}
\value{
a data frame with 5 columns
\describe{
    \item{container_id}{Unique id of the container that ran the iteration.}
    \item{iteration_id}{Unique id of the iteration. See also the function 'generate_scenarios' in the sleepsimR library <https://github.com/JasperHG90/sleepsimR>}
    \item{ts_request}{POSIXct. Date/time stamp indicating when the iteration started.}
    \item{ts_finished}{POSIXct. Date/time stamp indicating when the iteration finished.}
    \item{runtime_minutes}{Numeric. Time, in minutes, between ts_request and ts_finished.}
}
}
\description{
Convert allocations to data frame
}
