% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convergence.R
\name{autocorr_plot}
\alias{autocorr_plot}
\title{Make autocorrelation plots of the convergence of two chains}
\usage{
autocorr_plot(mod, param, var, thin = 5)
}
\arguments{
\item{param}{character. name of the variable to plot.}

\item{var}{numeric. index of the emission distribution.}

\item{thin}{numeric. thinning interval for the posterior distribution. Defaults to 5.}

\item{mod1}{\link[mHMMbayes]{mHMM_cont} model.}

\item{mod2}{\link[mHMMbayes]{mHMM_cont} model.}
}
\description{
Make autocorrelation plots of the convergence of two chains
}
\seealso{
Explanation in \link[coda]{autocorr.plot}.
}
