% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{average_runtime}
\alias{average_runtime}
\title{Compute the average runtime of the allocations}
\usage{
average_runtime(x, ...)
}
\arguments{
\item{x}{object of type sleepsimR_allocations}
}
\value{
numeric scalar. Average runtime across all iterations.
}
\description{
Compute the average runtime of the allocations
}
