% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{bias_MCMC_SE}
\alias{bias_MCMC_SE}
\title{MCMC Standard Error of the bias value}
\usage{
bias_MCMC_SE(x)
}
\arguments{
\item{x}{k-length numeric vector. k >= 1 and holds the parameter values of the estimated parameters.}
}
\value{
Bias MCMC SE
}
\description{
MCMC Standard Error of the bias value
}
