% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convergence.R
\name{combine_mcmc_chains}
\alias{combine_mcmc_chains}
\title{Combine two separate chains into one chain}
\usage{
combine_mcmc_chains(mod1, mod2, param, var, thin = 5)
}
\description{
Combine two separate chains into one chain
}
