% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{combine_posterior_chains}
\alias{combine_posterior_chains}
\title{Combine the posterior distributions of two models into one posterior distribution.}
\usage{
combine_posterior_chains(mod_res)
}
\arguments{
\item{mod_res}{list containing two \link[mHMMbayes]{mHMM_cont} models.}
}
\value{
List containing combined posterior distribution for the following variables:
\itemize{
    \item{emiss_mu_bar}{between-subject means of the emission distributions.}
    \item{emiss_varmu_bar}{between-subject variances of the emission distributions.}
    \item{emiss_var_bar}{residual variances of the emission distributions. (assumed fixed across individuals).}
    \item{gamma_int_bar}{intercept of the linear predictors for the transition probabilities.}
    \item{gamma_prob_bar}{gamma_int_bar transformed by the multinomial logistic regression formula.}
    \item{PD_subj}{subject-specific state-dependent emission distribution means for each subject.}
    \item{gamma_int_subj}{subject-specific intercepts of the linear predictor for the transition probabilities.}
    \item{gamma_prob_subj}{gamma_int_subj transformed by the multinomial logistic regresion formula.}
}
}
\description{
This function is used to combine the posterior distributions of two models.
}
