% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{compute_state_length}
\alias{compute_state_length}
\title{Compute the state duration for each state across three equal-size splits of the input data.}
\usage{
compute_state_length(input_data)
}
\description{
Compute the state duration for each state across three equal-size splits of the input data.
}
