% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{compute_tpm}
\alias{compute_tpm}
\title{Compute transition probabilities from observed data set.}
\usage{
compute_tpm(input_data)
}
\description{
Compute transition probabilities from observed data set.
}
