% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{coverage_MCMC_SE}
\alias{coverage_MCMC_SE}
\title{Coverage MCMC Standard Error}
\usage{
coverage_MCMC_SE(coverage, nsim)
}
\arguments{
\item{coverage}{XX}

\item{nsim}{XX}
}
\value{
XX
}
\description{
Coverage MCMC Standard Error
}
