% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convergence.R
\name{dens_plot}
\alias{dens_plot}
\title{Make density plots of the convergence of two chains}
\usage{
dens_plot(mod1, mod2, param, var, thin = 5)
}
\arguments{
\item{mod1}{\link[mHMMbayes]{mHMM_cont} model.}

\item{mod2}{\link[mHMMbayes]{mHMM_cont} model.}

\item{param}{character. name of the variable to plot.}

\item{var}{numeric. index of the emission distribution.}

\item{thin}{numeric. thinning interval for the posterior distribution. Defaults to 5.}
}
\description{
Make density plots of the convergence of two chains
}
\seealso{
Explanation in \link[mcmcplots]{denoverplot}.
}
