% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{emperical_MCMC_SE}
\alias{emperical_MCMC_SE}
\title{Compute the MCMC SE}
\usage{
emperical_MCMC_SE(emp_se, nsim)
}
\arguments{
\item{x}{numeric vector. Simulated parameter estimates.}
}
\value{
numeric scalar. MCMC standard error.
}
\description{
Compute the MCMC SE
}
