% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{MSE_MCMC_SE}
\alias{MSE_MCMC_SE}
\title{MSE MCMC Standard Error}
\usage{
MSE_MCMC_SE(MSE, estimates, true_value, nsim)
}
\arguments{
\item{nsim}{}
}
\value{
YY
}
\description{
MSE MCMC Standard Error
}
