% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{parse_sleepsimR_results}
\alias{parse_sleepsimR_results}
\title{Parse multiple sleepsimR result files}
\usage{
parse_sleepsimR_results(folder_path)
}
\arguments{
\item{folder_path}{}
}
\value{

}
\description{
Parse multiple sleepsimR result files
}
