% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{postprocess_gamma_int}
\alias{postprocess_gamma_int}
\title{Postprocess utility function for gamma_int_bar}
\usage{
postprocess_gamma_int(z, m)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Postprocess utility function for gamma_int_bar
}
