% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{postprocess_param_est}
\alias{postprocess_param_est}
\title{Postprocessing utility function for parameter estimates}
\usage{
postprocess_param_est(z, m)
}
\arguments{
\item{m}{integer. Number of hidden states}
}
\value{

}
\description{
Postprocessing utility function for parameter estimates
}
