% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{postprocess_subject_specific}
\alias{postprocess_subject_specific}
\title{Postprocess subject-specific metrics}
\usage{
postprocess_subject_specific(z, m)
}
\arguments{
\item{m}{}
}
\value{

}
\description{
Postprocess subject-specific metrics
}
