% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{PPP_mean_var}
\alias{PPP_mean_var}
\title{Posterior predictive check on emission means and variances.}
\usage{
PPP_mean_var(newdata)
}
\arguments{
\item{newdata}{simulated dataset created with \link[sleepsimReval]{simulate_new_data}.}
}
\description{
Posterior predictive check on emission means and variances.
}
\seealso{
section XX.YY in my thesis.
}
