% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{PPP_tpm}
\alias{PPP_tpm}
\title{Posterior predictive check on transition probabilities.}
\usage{
PPP_tpm(newdata)
}
\arguments{
\item{newdata}{simulated dataset created with \link[sleepsimReval]{simulate_new_data}.}
}
\description{
Posterior predictive check on transition probabilities.
}
\seealso{
section XX.YY in my thesis.
}
