% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocations.R
\name{read_allocations_file}
\alias{read_allocations_file}
\title{Read a sleepsimR allocations file}
\usage{
read_allocations_file(path)
}
\arguments{
\item{path}{path to the allocations file}
}
\value{
S3 object of type 'sleepsimR_allocations'
}
\description{
Read a sleepsimR allocations file
}
\details{
A sleepsimR allocations file is a json file that has been created
    by the sleepsimR-api program <https://github.com/JasperHG90/sleepsimR-api>.
    It contains two elements: (1) a list of dicts where the key is the container-id
    (i.e. the container that processed an iteration of the study) and the values
    contain (a) the unique id of the iteration on which the program worked, (b)
    the status of the program ('working', 'error' or 'completed') and (c) two timestamp
    values (if completed). The first time stamp value indicates when the iteration
    began, and the second timestamp value indicates when it was finished. This function
    only returns containers which completed their task.
}
