% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{simulate_dataset_PPP}
\alias{simulate_dataset_PPP}
\title{This is an internal function that generates a new dataset given the inputs}
\usage{
simulate_dataset_PPP(n, n_t, gamma, mu, eps, zeta, Q, seed)
}
\arguments{
\item{n}{int. number of subjects.}

\item{n_t}{int. number of observations for each subject.}

\item{gamma}{vector of length m times m containing transition probabilities.}

\item{mu}{vector of length m times m with emission distribution means.}

\item{eps}{vector of length m times m with residual variances.}

\item{zeta}{vector of floats. between-subject variance of the emission distributions.}

\item{Q}{float. between-subject variance of the transition probability matrix.}

\item{seed}{int. random seed used to generate the dataset.}
}
\value{
simulated dataset. See \link[mHMMbayes]{sim_mHMM}
}
\description{
This is an internal function that generates a new dataset given the inputs
}
