% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_PPP.R
\name{simulate_new_data}
\alias{simulate_new_data}
\title{Simulate a sleep dataset used in posterior predictive checks.}
\usage{
simulate_new_data(posterior)
}
\arguments{
\item{posterior}{list of posterior distributions created by the function \link[sleepsimReval]{combine_posterior_chains}.}
}
\value{
list containing states and emission distribution observed data. See \link[mHMMbayes]{sim_mHMM}.
}
\description{
Simulate a sleep dataset used in posterior predictive checks.
}
