% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_ggtheme.R
\name{theme_thesis}
\alias{theme_thesis}
\title{Custom theme for thesis plots}
\usage{
theme_thesis(
  text_size = 12,
  line_size = text_size/170,
  rect_size = text_size/170,
  center_title = FALSE
)
}
\arguments{
\item{text_size}{text size for thesis plots}

\item{line_size}{size of the axis lines}

\item{rect_size}{sizs of the background lines}
}
\value{
adds thesis theme to a plot generated by ggplot2
}
\description{
Custom theme for thesis plots
}
\seealso{
Adapted from https://www.r-bloggers.com/custom-themes-in-ggplot2/
}
