% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_convergence.R
\name{tpp}
\alias{tpp}
\title{Make trace plots of the convergence of two chains}
\usage{
tpp(mod1, mod2, param, var, thin = 5)
}
\arguments{
\item{mod1}{\link[mHMMbayes]{mHMM_cont} model.}

\item{mod2}{\link[mHMMbayes]{mHMM_cont} model.}

\item{param}{character. name of the variable to plot.}

\item{var}{numeric. index of the emission distribution.}

\item{thin}{numeric. thinning interval for the posterior distribution. Defaults to 5.}
}
\description{
Make trace plots of the convergence of two chains
}
\seealso{
Explanation in \link[coda]{traceplot}.
}
