% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_metrics.R
\name{emperical_SE}
\alias{emperical_SE}
\title{Compute the emperical SE}
\usage{
emperical_SE(x)
}
\arguments{
\item{x}{numeric vector. Simulated parameter estimates}
}
\value{
numeric vector with two values: (1) Emperical standard error and (2) MCMC SE of this value
}
\description{
Compute the emperical SE
}
\seealso{
Morris, Tim P., Ian R. White, and Michael J. Crowther. "Using simulation studies to evaluate statistical methods." Statistics in medicine 38.11 (2019): 2074-2102.
}
