% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{parse_sleepsimR_result}
\alias{parse_sleepsimR_result}
\title{Parse a sleepsimR result file}
\usage{
parse_sleepsimR_result(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Parse a sleepsimR result file
}
